requirements.txt
# Core
python-dotenv==1.0.1
openai==1.8.3

# LangChain
langchain==0.1.242
langchain-openai==0.0.17

# Gradio stable versions
gradio==4.49.0
gradio-client==0.4.0

# Optional
requests==2.31.0

